This package is intended to be a disk-instalable package, it doesn't use libsmbcommon, and can load external modules. It's binaries are huge.
By contrary, in the samba-small package libsmbcommon.so is created and used by all
samba binaries (smbd nmbd smbtree smbstatus swat smbpasswd), significaly
reducing its size. This is intended to be incorporated in the firmware,
given the low amount of available flash space. But this removes the
hability to load modules.




git-svn-id: https://svn.code.sf.net/p/alt-f/code/trunk/alt-f@2411 cd80dc97-fa67-4d1f-9ead-3176a861659d